requirements.txt
alembic==1.8.1
amqp==5.1.1
anyio==3.6.1
apache-airflow==2.3.4
apache-airflow-providers-celery==3.0.0
apache-airflow-providers-common-sql==1.1.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.2.0
apispec==3.3.2
argcomplete==2.0.0
async-generator==1.10
attrs==22.1.0
Babel==2.10.3
billiard==3.6.4.0
blinker==1.5
cachelib==0.9.0
cattrs==22.1.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.5
colorlog==4.8.0
commonmark==0.9.1
connexion==2.14.1
cron-descriptor==1.2.31
croniter==1.3.5
cryptography==37.0.4
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
docutils==0.19
email-validator==1.2.1
exceptiongroup==1.0.0rc9
Flask==2.2.2
Flask-Admin==1.6.0
Flask-AppBuilder==4.1.3
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flower==1.2.0
graphviz==0.20.1
greenlet==1.1.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
humanize==4.3.0
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.15.0
kombu==5.2.4
lazy-object-proxy==1.7.1
linkify-it-py==2.0.0
lockfile==0.12.2
lxml==4.9.1
Mako==1.2.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.0
mdurl==0.1.2
numpy==1.23.2
outcome==1.2.0
packaging==21.3
pandas==1.4.4
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pyodbc==4.0.34
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-daemon==2.3.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.2.1
pytzdata==2020.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.5.1
selenium==4.4.3
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.40
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.2
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tornado==6.2
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.3.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.8.3
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
WTForms==2.3.3
zipp==3.8.1
